+++ src/main/java/com/project/coalba/domain/auth/info/impl/dto/NaverUserInfoDto.java
@@ -1,4 +1,4 @@
-package com.project.coalba.domain.auth.info.dto;
+package com.project.coalba.domain.auth.info.impl.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
